convert_music: Find foreign files and empty dirs.
find_files() now includes list of foreign (non-mp3) files in the mp3
dir. This will be used to notify the user since this isn't expected.

Added find_empty_dirs() to find empty directories (or directories of
empty directories) in the mp3 dir. This will be run at the end of
convert_music since it was too much work for me to add prediction logic
to find_files().

Also added styling tweaks and started working on the ConvertFiles class.